chore(api): rework react query useMutations
